+++ GenieWin8/GenieWin8/BypassAccountLogoutPage.xaml.cs
@@ -13,6 +13,7 @@ using Windows.UI.Xaml.Media;
 using Windows.UI.Xaml.Navigation;
 using GenieWin8.DataModel;
 using Windows.Storage;
+using Windows.Networking.Connectivity;
 
 // “基本页”项模板在 http://go.microsoft.com/fwlink/?LinkId=234237 上有介绍
 
@@ -23,9 +24,34 @@ namespace GenieWin8
     /// </summary>
     public sealed partial class BypassAccountLogoutPage : GenieWin8.Common.LayoutAwarePage
     {
+        private static bool IsWifiSsidChanged;
         public BypassAccountLogoutPage()
         {
             this.InitializeComponent();
+            Application.Current.Resuming += new EventHandler<Object>(App_Resuming);
+        }
+
+        private void App_Resuming(Object sender, Object e)
+        {
+            //判断所连接Wifi的Ssid是否改变
+            IsWifiSsidChanged = true;
+            try
+            {
+                var ConnectionProfiles = NetworkInformation.GetConnectionProfiles();
+                foreach (var connectionProfile in ConnectionProfiles)
+                {
+                    if (connectionProfile.GetNetworkConnectivityLevel() != NetworkConnectivityLevel.None)
+                    {
+                        if (connectionProfile.ProfileName == MainPageInfo.ssid)
+                            IsWifiSsidChanged = false;
+                        else
+                            IsWifiSsidChanged = true;
+                    }
+                }
+            }
+            catch (Exception ex)
+            {
+            }
         }
 
         /// <summary>
@@ -55,22 +81,30 @@ namespace GenieWin8
 
         private async void LogoutButton_Click(Object sender, RoutedEventArgs e)
         {
-            InProgress.IsActive = true;
-            PopupBackgroundTop.Visibility = Visibility.Visible;
-            PopupBackground.Visibility = Visibility.Visible;
-            GenieSoapApi soapApi = new GenieSoapApi();
-            Dictionary<string, string> dicResponse = new Dictionary<string, string>();
-            UtilityTool util = new UtilityTool ();
-            string MacAddress = util.GetLocalMacAddress();  //获取本机mac地址
-            MacAddress = MacAddress.Replace(":","");
-            dicResponse = await soapApi.DeleteMACAddress(MacAddress);
-            ParentalControlInfo.BypassUsername = "";
-            ParentalControlInfo.BypassChildrenDeviceId = "";
-            WriteChildrenDeviceIdToFile();                  //登录成功后将childrenDeviceId保存到本地，如果未注销则以后登录Genie时，通过读取本地DeviceId获得当前登录的Bypass账户
-            InProgress.IsActive = false;
-            PopupBackgroundTop.Visibility = Visibility.Collapsed;
-            PopupBackground.Visibility = Visibility.Collapsed;
-            this.Frame.Navigate(typeof(ParentalControlPage));
+            if (IsWifiSsidChanged)
+            {
+                this.Frame.Navigate(typeof(LoginPage));
+                MainPageInfo.navigatedPage = "ParentalControlPage";
+            } 
+            else
+            {
+                InProgress.IsActive = true;
+                PopupBackgroundTop.Visibility = Visibility.Visible;
+                PopupBackground.Visibility = Visibility.Visible;
+                GenieSoapApi soapApi = new GenieSoapApi();
+                Dictionary<string, string> dicResponse = new Dictionary<string, string>();
+                UtilityTool util = new UtilityTool();
+                string MacAddress = util.GetLocalMacAddress();  //获取本机mac地址
+                MacAddress = MacAddress.Replace(":", "");
+                dicResponse = await soapApi.DeleteMACAddress(MacAddress);
+                ParentalControlInfo.BypassUsername = "";
+                ParentalControlInfo.BypassChildrenDeviceId = "";
+                WriteChildrenDeviceIdToFile();                  //登录成功后将childrenDeviceId保存到本地，如果未注销则以后登录Genie时，通过读取本地DeviceId获得当前登录的Bypass账户
+                InProgress.IsActive = false;
+                PopupBackgroundTop.Visibility = Visibility.Collapsed;
+                PopupBackground.Visibility = Visibility.Collapsed;
+                this.Frame.Navigate(typeof(ParentalControlPage));
+            }
         }
 
         public async void WriteChildrenDeviceIdToFile()
